+++ src/components/content/conditions/unread-conditions-notification.tsx
@@ -1,0 +1,11 @@
+import { UnreadNotification } from "@/components/core/unread-notification";
+import { usePatientConditionsOutside } from "@/services/conditions";
+
+export type UnreadConditionsNotificationProps = {
+  className?: string;
+};
+
+export const UnreadConditionsNotification = ({ className }: UnreadConditionsNotificationProps) => {
+  const query = usePatientConditionsOutside();
+  return <UnreadNotification className={className} query={query} />;
+};