+++ src/components/content/document/helpers/columns.tsx
@@ -3,21 +3,27 @@ import { DocumentModel } from "@/fhir/models/document";
 
 export const patientDocumentColumns: TableColumn<DocumentModel>[] = [
   {
-    widthPercent: 20,
-    minWidth: 100,
-    title: "Date Created",
-    render: (document) => <div className="group-hover:ctw-underline">{document.dateCreated}</div>,
-  },
-  {
-    widthPercent: 30,
+    widthPercent: 35,
     minWidth: 200,
     title: "Title",
-    render: (document) => <div>{document.title}</div>,
+    render: (document) => <div className="group-hover:ctw-underline">{document.title}</div>,
   },
   {
-    widthPercent: 30,
+    widthPercent: 15,
+    minWidth: 128,
+    title: "Encounter Date",
+    dataIndex: "encounterDate",
+  },
+  {
+    widthPercent: 15,
+    minWidth: 128,
+    title: "Date Retrieved",
+    dataIndex: "dateCreated",
+  },
+  {
+    widthPercent: 35,
     minWidth: 200,
-    title: "Managing Organization",
+    title: "Author",
     render: (document) => <div>{document.custodian}</div>,
   },
 ];