+++ src/components/content/document/patient-documents.tsx
@@ -1,6 +1,7 @@
 import cx from "classnames";
 import { useMemo, useRef } from "react";
 import { patientDocumentColumns } from "./helpers/columns";
+import { defaultDocumentSort, documentSortOptions } from "./helpers/sorts";
 import { getDateRangeView } from "../resource/helpers/view-date-range";
 import { useResourceDetailsDrawer } from "../resource/resource-details-drawer";
 import { ResourceTable } from "../resource/resource-table";
@@ -27,17 +28,19 @@ function PatientDocumentsComponent({ className, onAddToRecord }: PatientDocument
 
   const patientDocumentQuery = usePatientDocuments();
   const rowActions = useMemo(() => getRowActions({ onAddToRecord }), [onAddToRecord]);
-  const { viewOptions, defaultView } = getDateRangeView<DocumentModel>("dateCreated");
-  const { data, setViewOption } = useFilteredSortedData({
-    defaultView,
+  const { viewOptions, allTime } = getDateRangeView<DocumentModel>("dateCreated");
+  const { data, setViewOption, setSort } = useFilteredSortedData({
+    defaultSort: defaultDocumentSort,
+    defaultView: allTime,
     records: patientDocumentQuery.data,
   });
 
-  const isEmptyQuery = patientDocumentQuery.data?.length === 0;
+  const isEmptyQuery = patientDocumentQuery.data.length === 0;
   const hasZeroFilteredRecords = !isEmptyQuery && data.length === 0;
 
   const openDetails = useResourceDetailsDrawer({
-    header: (m) => `${m.dateCreated} - ${m.title}`,
+    header: (m) => m.title,
+    subHeader: (m) => m.encounterDate,
     details: documentData,
     enableFQS: enabled,
   });
@@ -49,10 +52,15 @@ function PatientDocumentsComponent({ className, onAddToRecord }: PatientDocument
       className={cx(className, "ctw-scrollable-pass-through-height")}
     >
       <ResourceTableActions
+        sortOptions={{
+          defaultSort: defaultDocumentSort,
+          options: documentSortOptions,
+          onChange: setSort,
+        }}
         viewOptions={{
           onChange: setViewOption,
           options: viewOptions,
-          defaultView,
+          defaultView: allTime,
         }}
       />
       <ResourceTable
@@ -64,6 +72,7 @@ function PatientDocumentsComponent({ className, onAddToRecord }: PatientDocument
         columns={patientDocumentColumns}
         onRowClick={openDetails}
         RowActions={rowActions}
+        enableDismissAndReadActions
       />
     </div>
   );
@@ -106,18 +115,17 @@ const RowActions = ({ record, onAddToRecord }: RowActionsProps2) => {
 export const PatientDocuments = withErrorBoundary(PatientDocumentsComponent, "PatientDocuments");
 
 const documentData = (document: DocumentModel) => [
-  { label: "status", value: document.status },
-  { label: "docStatus", value: document.docStatus },
-  { label: "Managing Organization", value: document.custodian },
+  { label: "Date Retrieved", value: document.dateCreated },
+  { label: "Author", value: document.custodian },
   {
     label: "Section Display",
     value: document.sectionDisplays && (
-      <div>
+      <ul className="ctw-m-0 ctw-pl-4">
         {document.sectionDisplays.map((item, index) => (
           // eslint-disable-next-line react/no-array-index-key
-          <div key={item + index}>{item}</div>
+          <li key={item + index}>{item}</li>
         ))}
-      </div>
+      </ul>
     ),
   },
 ];